invite link and admin checks

+++ utils/api.ts
@@ -46,10 +46,7 @@ class API {
   //////////// AUTH REQUESTS ////////////
   ///////////////////////////////////////
 
-  public async login(
-    username: string,
-    password: string,
-  ): Promise<LoginResponse> {
+  public async login(username: string, password: string): Promise<LoginResponse> {
     try {
       const response = await this.client.post<LoginResponse>("/login", {
         username,
@@ -106,14 +103,9 @@ class API {
     }
   }
 
-  public async postNewUserData(
-    newUserData: UserMetadata,
-  ): Promise<UserMetadata> {
+  public async postNewUserData(newUserData: UserMetadata): Promise<UserMetadata> {
     try {
-      const response = await this.client.post<UserMetadata>(
-        "/protected/user/meta",
-        newUserData,
-      );
+      const response = await this.client.post<UserMetadata>("/protected/user/meta", newUserData);
       console.log("User data modification successful");
       return response.data;
     } catch (error) {
@@ -129,14 +121,9 @@ class API {
   //////////////////////////////////////////
   //////////// CHANNEL REQUESTS ////////////
   //////////////////////////////////////////
-  public async createNewChannel(
-    newChannelData: NewChannelData,
-  ): Promise<number> {
+  public async createNewChannel(newChannelData: NewChannelData): Promise<number> {
     try {
-      const response = await this.client.post<number>(
-        "/protected/channel",
-        newChannelData,
-      );
+      const response = await this.client.post<number>("/protected/channel", newChannelData);
       console.log("Channel creation successful");
       return response.data;
     } catch (error) {
@@ -160,18 +147,13 @@ class API {
     }
   }
 
-  public async addUserToChannel(
-    channelId: number,
-    userId: string,
-  ): Promise<string> {
+  public async addUserToChannel(channelId: number, userId: string): Promise<string> {
     try {
-      await this.client.put(
-        `/protected/channel/${channelId}/user/${userId}`,
-        {},
-      );
+      await this.client.put(`/protected/channel/${channelId}/user/${userId}`, {});
       console.log("User added to channel");
       return "added";
     } catch (error) {
+      console.error(error);
       if (axios.isAxiosError(error) && error.response?.status === 401) {
         throw new Error("Can't add user to channel, invalid token");
       }
@@ -179,14 +161,9 @@ class API {
     }
   }
 
-  public async banUserFromChannel(
-    channelId: number,
-    userId: string,
-  ): Promise<string> {
+  public async banUserFromChannel(channelId: number, userId: string): Promise<string> {
     try {
-      await this.client.delete(
-        `/protected/channel/${channelId}/user/${userId}`,
-      );
+      await this.client.delete(`/protected/channel/${channelId}/user/${userId}`);
       console.log("User removed from channel");
       return "removed";
     } catch (error) {
@@ -197,22 +174,14 @@ class API {
     }
   }
 
-  public async updateChannel(
-    channelId: number,
-    newChannelData: ChannelUpdateMetadata,
-  ): Promise<string> {
+  public async updateChannel(channelId: number, newChannelData: ChannelUpdateMetadata): Promise<string> {
     try {
-      await this.client.put(
-        `/protected/channel/${channelId}/update_metadata`,
-        newChannelData,
-      );
+      await this.client.put(`/protected/channel/${channelId}/update_metadata`, newChannelData);
       console.log("Channel updated");
       return "updated";
     } catch (error) {
       if (axios.isAxiosError(error) && error.response?.status === 401) {
-        throw new Error(
-          "Can't update channel, you don't have permissions to perform this action",
-        );
+        throw new Error("Can't update channel, you don't have permissions to perform this action");
       }
       throw new Error("Can't update channel");
     }
@@ -220,9 +189,7 @@ class API {
 
   public async getChannels(): Promise<ChannelMetadata[]> {
     try {
-      const response = await this.client.get<ChannelMetadata[]>(
-        `/protected/user/channels`,
-      );
+      const response = await this.client.get<ChannelMetadata[]>(`/protected/user/channels`);
       console.log("Channels retrieved");
       return response.data;
     } catch (error) {
@@ -233,18 +200,14 @@ class API {
     }
   }
 
-  public async createInvite(channelId: number): Promise<String> {
+  public async createInvite(channelId: number): Promise<string> {
     try {
-      const response = await this.client.post(
-        `/protected/channel/${channelId}/invite_link`,
-      );
+      const response = await this.client.post(`/protected/channel/${channelId}/invite_link`);
       console.log("Invite link created");
       return response.data;
     } catch (error) {
       if (axios.isAxiosError(error) && error.response?.status === 401) {
-        throw new Error(
-          "You don't have permission to create invite links for this channel",
-        );
+        throw new Error("You don't have permission to create invite links for this channel");
       } else if (axios.isAxiosError(error) && error.response?.status === 404) {
         throw new Error("Channel not found");
       } else if (axios.isAxiosError(error) && error.response?.status === 500) {
@@ -277,10 +240,7 @@ class API {
   //////////// MESSAGES REQUESTS ////////////
   ///////////////////////////////////////////
 
-  public async getMessages(
-    channelId: number,
-    batchOffset: number,
-  ): Promise<MessageMetadata[]> {
+  public async getMessages(channelId: number, batchOffset: number): Promise<MessageMetadata[]> {
     try {
       const response = await this.client.get<MessageMetadata[]>(
         `/protected/channel/${channelId}/messages/${batchOffset}`,
@@ -290,51 +250,33 @@ class API {
       return response.data;
     } catch (error) {
       if (axios.isAxiosError(error) && error.response?.status === 401) {
-        throw new Error(
-          "Can't get messages, user does not have permission to use this channel",
-        );
+        throw new Error("Can't get messages, user does not have permission to use this channel");
       }
       throw new Error("Can't get messages");
     }
   }
 
-  public async sendMessage(
-    channelId: number,
-    newMessage: NewMessageData,
-  ): Promise<string> {
+  public async sendMessage(channelId: number, newMessage: NewMessageData): Promise<string> {
     try {
-      await this.client.post(
-        `/protected/channel/${channelId}/message`,
-        newMessage,
-      );
+      await this.client.post(`/protected/channel/${channelId}/message`, newMessage);
       console.log("Message sent");
       return "sent";
     } catch (error) {
       if (axios.isAxiosError(error) && error.response?.status === 401) {
-        throw new Error(
-          "Can't send message, user does not have permission to use this channel",
-        );
+        throw new Error("Can't send message, user does not have permission to use this channel");
       }
       throw new Error("Can't send message: " + error);
     }
   }
 
-  public async updateMessage(
-    channelId: number,
-    messageUpdate: MessageMetadata,
-  ): Promise<string> {
+  public async updateMessage(channelId: number, messageUpdate: MessageMetadata): Promise<string> {
     try {
-      await this.client.post(
-        `/protected/channel/${channelId}/message/moderate`,
-        messageUpdate,
-      );
+      await this.client.post(`/protected/channel/${channelId}/message/moderate`, messageUpdate);
       console.log("Message updated");
       return "updated";
     } catch (error) {
       if (axios.isAxiosError(error) && error.response?.status === 401) {
-        throw new Error(
-          "Can't moderate message, user does not have permission to use this channel",
-        );
+        throw new Error("Can't moderate message, user does not have permission to use this channel");
       }
       throw new Error("Can't moderate message");
     }
